+++ lua/CompMod/Alien/Drifters/Abilities/Hallucination/DrifterHeight/Post/Hallucination.lua
@@ -1,0 +1,11 @@
+kHallucinationHoverHeight = 1
+
+if Server then
+    function Hallucination:GetHoverHeight()
+        if self.assignedTechId == kTechId.Lerk or self.assignedTechId == kTechId.Drifter then
+            return kHallucinationHoverHeight
+        else
+            return 0
+        end
+    end
+end