+++ docs/changes.md
@@ -41,6 +41,7 @@
     * Charge
         * Knockback removed
 * Drifters
+    * Reduced hover height to 1 from 1.2 so they don't float in marine's faces
     * Abilities
         * Mucous
             * Reduced cloud radius from 8 to 5 (reduces area by 60%)
@@ -49,6 +50,7 @@
             * No longer affected by mucous
             * Can no longer have babblers attached
             * Reduced Onos HP to 100 eHP
+            * Reduced hallucination drifter's hover height to match real drifter
 * Structure Abilities
     * Echo
         * Echo upgrade cost increased to 2 from 1